api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai-tools>=0.14.0",
    "crewai>=0.80.0",
    "fastapi>=0.115.5",
    "humanlayer==0.6.1",
    "pydantic==2.8.2",
    "python-dotenv>=1.0.1",
    "stripe>=11.2.0",
    "stripe-agent-toolkit>=0.1.21",
    "uvicorn[standard]>=0.32.0",
]
